feat: move api requests to thunks
