+++ src/pages/form/FormPage.tsx
@@ -7,7 +7,7 @@ import { TUserCardModel } from '../../types';
 import style from './FormPage.module.scss';
 
 export const FormPage = () => {
-  const { formValues } = useAppSelector((state) => state.search);
+  const { formValues } = useAppSelector((state) => state.form);
   const dispatch = useAppDispatch();
 
   const setValues = (user: TUserCardModel) => {
